gluon-alt-esc: add client + provider package
The client package allows to create a new wifi interface with direct access
to the local Freifunk network but with alternative exit gateways for internet
connectivity. A custom ESSID can be chosen.

The provider package is the counterpart to the client package and
configures the firewall of the selected Gluon node to grant permission to
route packets between the client and wan zone.

Note that the Gluon Alt-ESC provider package is not mandatory for the
Gluon Alt-ESC client package. In fact, any client device in the mesh network
can be chosen and configured to provide internet access for the Alt-ESC
client package.

Known issues + remaining tasks:
* Using Alt-ESC client + provider on the same node does not work yet
  (looks like a kernel bug between ebtables-dnat and IP routing code)
* Documentation++
* Translations
* Add VLAN tag configuration option to the advanced settings of the
  mesh-vpn package to make VPN traffic distinguashable from traffic
  routed via Gluon Alt-ESC
  (would be useful for middleboxes to forward mesh-VPN traffic
   directly, but Gluon ALT-ESC traffic over a VPN established by the
   middlebox, possibly an anonymizing VPN)
* Maybe technically disallow the word "Freifunk" in the custom ESSID?