Drop accumulated survey dumps for ath10k
The QCA988X firmware seems to hang all the time when gathering the survey
dumps. It is currently not possible to gather meaningful data on these
chips without extreme high cpu load and hanging processes.

Just drop the support at all and switch to another workaround at a later
point.